% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_vuepress.R
\name{install_vuepress}
\alias{install_vuepress}
\title{Install VuePress}
\usage{
install_vuepress(global = FALSE, dir = ".")
}
\arguments{
\item{global}{If \code{TRUE}, the VuePress install to Node.js global package path.
In other words, identical to \code{yarn global add}.}

\item{dir}{The directory to install VuePress.}
}
\description{
Download the VuePress executable with Node.js package manager.
You can select install path that global or specific directory.
}
\examples{
\dontrun{
install_vuepress(global = TRUE)
install_vuepress(dir = "project")
}
}
