% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_site}
\alias{render_site}
\title{Render multiple R Markdown files convert to Markdown}
\usage{
render_site(input_dir = ".")
}
\arguments{
\item{input_dir}{Directory that target
R Markdown to convert into Markdown.}
}
\description{
Render multiple R Markdown files convert to Markdown
}
\examples{
\dontrun{
render_site("project")
}
}
